refactor: divide ciphers into separate files, add gitignore, add jest as dev dependency
