+++ lib/ciphers/constants.js
@@ -1,0 +1,13 @@
+const ALPHABET = 'abcdefghijklmnopqrstuvwxyz'.split('');
+const REVERSE_ALPHABET = ALPHABET.reverse();
+const DECODING_MODE = 'decoding';
+const ENCODING_MODE = 'encoding';
+const ENCRIPTION_MODE_CODES = [DECODING_MODE, ENCODING_MODE];
+
+module.exports = {
+    ALPHABET,
+    REVERSE_ALPHABET,
+    DECODING_MODE,
+    ENCODING_MODE,
+    ENCRIPTION_MODE_CODES,
+};
