+++ lib/cliOptionsParser.js
@@ -1,4 +1,4 @@
-const fs = require('fs');
+const { access, constants } = require('fs');
 const { splitChunks } = require('./utils');
 const handleError = require('./errorHandler');
 const {
@@ -72,7 +72,7 @@ const isDublicateKey = (key) => {
 }
 
 const validateInputPath = (path) => {
-    fs.access(path, fs.constants.R_OK | fs.constants.F_OK, (err) => {
+    access(path, constants.R_OK | constants.F_OK, (err) => {
         if (err) {
             handleError(`Input '${path}' is not accessable`)
             return;
@@ -81,7 +81,7 @@ const validateInputPath = (path) => {
 }
 
 const validateOutputPath = (path) => {
-    fs.access(path, fs.constants.W_OK | fs.constants.F_OK, (err) => {
+    access(path, constants.W_OK | constants.F_OK, (err) => {
         if (err) {
             handleError(`Output '${path}' is not accessable`)
             return;
